gcloud/functions/helper/requirements.txt
google-cloud-pubsub<3
google-cloud-logging